test: put test in right spot (#902)
